# openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "ciberpunk-storymaker",
  "version": "0.1.0"
 },
 "paths": {
...
  "/lectura/{token}": {
   "get": {
    "tags": [
     "lectura"
    ],
    "summary": "Portada",
    "description": "La portada: dedicatoria, indice y nada de prosa.\n\n**Los diez capitulos no vienen embebidos.** Serian unas doce mil palabras\ncargadas al abrir la pagina donde no se lee ni una; la lectura es capitulo\na capitulo y la peticion tambien.\n\nLa dedicatoria se sirve aqui, y es su unico uso: la creo T2 y hasta ahora\nno la leia nadie.",
    "operationId": "portada_lectura__token__get",
    "parameters": [
     {
      "name": "token",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VersionPublicada"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/lectura/{token}/capitulos/{numero}": {
   "get": {
    "tags": [
     "lectura"
    ],
    "summary": "Capitulo",
    "description": "La prosa de un capitulo, **la que se fijo al publicar**.\n\nNo la vigente: `capitulo_publicado` apunta a un `version_texto_id` concreto\n(`RF-PUB-01`), y resolver la version de hoy haria que un enlace ya\nrepartido cambiara de contenido sin que nadie lo tocara.",
    "operationId": "capitulo_lectura__token__capitulos__numero__get",
    "parameters": [
     {
      "name": "token",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Token"
      }
     },
     {
      "name": "numero",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Numero"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Capitulo Lectura  Token  Capitulos  Numero  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/lectura/{token}/ficha": {
   "get": {
    "tags": [
     "lectura"
    ],
    "summary": "Ficha",
    "description": "Quien es quien, y en que capitulos aparece (`RF-PUB-05`).\n\nUna tirada sin ficha devuelve la ficha vacia y no un 404: el enlace es\nvalido y la portada se lee igual. Un 404 aqui diria que el regalo no\nexiste, que es otra cosa.",
    "operationId": "ficha_lectura__token__ficha_get",
    "parameters": [
     {
      "name": "token",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FichaDeLectura"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/lectura/{token}/versiones": {
   "get": {
    "tags": [
     "lectura"
    ],
    "summary": "Versiones",
    "description": "Las tiradas de esta obra, de la mas reciente a la primera.\n\nSin prosa y sin capitulos: es el historial para volver a una ya leida, y\ncada una se abre por su propio enlace.",
    "operationId": "versiones_lectura__token__versiones_get",
    "parameters": [
     {
      "name": "token",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/VersionPublicada"
         },
         "title": "Response Versiones Lectura  Token  Versiones Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/lectura/{token}/pdf": {
   "get": {
    "tags": [
     "lectura"
    ],
    "summary": "Pdf",
    "description": "El regalo en un fichero, para imprimir o guardar (T10, `RI-11`).\n\n**Sale de la version publicada, nunca del texto vigente** (regla de dominio\n14): cada capitulo se lee con `texto_publicado`, lo mismo que la ruta del\ncapitulo. Y la dedicatoria va en la portada, no como capitulo (regla 15).\n\nRespuesta sincrona y no trabajo en segundo plano: el spike midio 1,2-2,4 s\npor novela, casi todo arranque del navegador.\n\nSi no hay navegador, **503 con motivo y no 500**. `ImpresoraNoDisponible`\nno es un error de dominio -- es una herramienta ausente -- y el manejador\ncentral de `commons/errors/` no la conoce; se traduce aqui, que es el unico\nsitio que la ve. **Ni el HTML ni la prosa se escriben en ningun log.**",
    "operationId": "pdf_lectura__token__pdf_get",
    "parameters": [
     {
      "name": "token",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "La novela publicada, en PDF",
      "content": {
       "application/pdf": {}
      }
     },
     "404": {
      "description": "No hay nada en este enlace."
     },
     "503": {
      "description": "No hay navegador con el que imprimir"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/obras/{obra_id}/publicar": {
   "post": {
    "tags": [
     "publicacion"
    ],
    "summary": "Publicar Obra",
    "description": "Fija una tirada inmutable y **devuelve el token con el que se lee**.\n\n`publicar` existia en el servicio desde la Fase 4 y no tenia ruta, asi que\ndesde el navegador no habia forma de publicar ni de saber cuando una novela\nse podia leer: el frontend llegaba hasta «se esta escribiendo» y ahi se\ncortaba. No faltaba una pantalla, **faltaba el dato**.\n\n**Es la unica ruta que devuelve el `identificador_publico`**, y la asimetria\ncon las de lectura es deliberada: alli no sale porque quien lee ya entro con\nel, y repetirlo solo lo pondria en un sitio mas -- un registro, una captura,\nun historial compartido --. Aqui sale porque quien publica aun no lo tiene.\n\nLos fallos de dominio -- un capitulo que no paso su puerta, una cronologia\nque Lean rechaza -- los traduce el manejador central de `commons/errors`\n(`CLAUDE.md` §6). Aqui no hay ni un `HTTPException`: un 500 dejaria al\nfrontend sin nada que contarle a quien acaba de encargar la novela.",
    "operationId": "publicar_obra_obras__obra_id__publicar_post",
    "parameters": [
     {
      "name": "obra_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Obra Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Publicar Obra Obras  Obra Id  Publicar Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/obras/{obra_id}/versiones/{version}/revertir": {
   "post": {
    "tags": [
     "publicacion"
    ],
    "summary": "Revertir La Version",
    "description": "RI-10. `version` es el identificador de la `VersionPublicada`.\n\n**No es larga y no va a segundo plano:** mover una bandera no llama al\nmodelo. La revertida no se borra (RF-PET-08) y su enlace sigue leyendose.",
    "operationId": "revertir_la_version_obras__obra_id__versiones__version__revertir_post",
    "parameters": [
     {
      "name": "obra_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Obra Id"
      }
     },
     {
      "name": "version",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Version"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Revertir La Version Obras  Obra Id  Versiones  Version  Revertir Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
...
   "CapituloPublicado": {
    "properties": {
     "numero": {
      "type": "integer",
      "title": "Numero"
     },
     "titulo": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Titulo"
     },
     "cambiado": {
      "type": "boolean",
      "title": "Cambiado",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "numero"
    ],
    "title": "CapituloPublicado",
    "description": "Una entrada del indice de la 002 (`RF-IND-01`, `RF-IND-02`).\n\n`cambiado` viene del backend calculado y no se recalcula en el navegador: el\nnavegador solo tiene delante la tirada que esta leyendo, y `cambiado`\ncompara con la anterior."
   },
...
   "EntradaDeFicha": {
    "properties": {
     "nombre": {
      "type": "string",
      "title": "Nombre"
     },
     "tipo": {
      "type": "string",
      "title": "Tipo"
     },
     "descripcion": {
      "type": "string",
      "title": "Descripcion",
      "default": ""
     },
     "capitulos": {
      "items": {
       "type": "integer"
      },
      "type": "array",
      "title": "Capitulos"
     },
     "hecho_canon_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Hecho Canon Id"
     }
    },
    "type": "object",
    "required": [
     "nombre",
     "tipo"
    ],
    "title": "EntradaDeFicha",
    "description": "Quien es quien, con los capitulos en que aparece (`RF-PUB-05`)."
   },
...
   "FichaDeLectura": {
    "properties": {
     "entradas": {
      "items": {
       "$ref": "#/components/schemas/EntradaDeFicha"
      },
      "type": "array",
      "title": "Entradas"
     }
    },
    "type": "object",
    "title": "FichaDeLectura",
    "description": "La ficha entera de una tirada: quien es quien, y donde sale.\n\nEs la **respuesta** de `/lectura/{token}/ficha`, no una fila: `EntradaDeFicha`\nes una entrada suelta y el frontend necesita el tipo del cuerpo completo\npara derivar su cliente."
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "VersionPublicada": {
    "properties": {
     "titulo": {
      "type": "string",
      "title": "Titulo",
      "default": ""
     },
     "ordinal": {
      "type": "integer",
      "title": "Ordinal"
     },
     "publicada_en": {
      "type": "string",
      "format": "date-time",
      "title": "Publicada En"
     },
     "dedicatoria": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Dedicatoria"
     },
     "capitulos": {
      "items": {
       "$ref": "#/components/schemas/CapituloPublicado"
      },
      "type": "array",
      "title": "Capitulos"
     }
    },
    "type": "object",
    "required": [
     "ordinal",
     "publicada_en"
    ],
    "title": "VersionPublicada",
    "description": "La tirada, **sin** `id` ni `obra_id`.\n\nEl `identificador_publico` no sale aqui: quien lee ya lo tiene, porque es\npor donde entro. Devolverlo otra vez solo lo pondria en un sitio mas -- un\nregistro, una captura, un historial del navegador compartido."
   }
  }
 }
}